refactor-as-shared-library
