+++ shisen_cpp/CMakeLists.txt
@@ -20,9 +20,14 @@ find_package(shisen_interfaces REQUIRED)
 
 install(DIRECTORY "include" DESTINATION ".")
 
-add_library(${PROJECT_NAME}
+add_library(${PROJECT_NAME} SHARED
   "src/capture_setting.cpp"
-  "src/interface.cpp")
+  "src/interface.cpp"
+  "src/consumer/capture_setting_consumer.cpp"
+  "src/consumer/image_consumer.cpp"
+  "src/node/camera_node.cpp"
+  "src/provider/capture_setting_provider.cpp"
+  "src/provider/image_provider.cpp")
 
 target_include_directories(${PROJECT_NAME} PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
