+++ shisen_cpp/include/shisen_cpp/provider/capture_setting_provider.hpp
@@ -38,15 +38,16 @@ public:
   {
   };
 
-  inline explicit CaptureSettingProvider(
+  explicit CaptureSettingProvider(
     rclcpp::Node::SharedPtr node, const Options & options = Options());
+  ~CaptureSettingProvider();
 
-  inline virtual CaptureSetting on_configure_capture_setting(
+  virtual CaptureSetting on_configure_capture_setting(
     const CaptureSetting & capture_setting);
 
-  inline void configure_capture_setting(const CaptureSetting & capture_setting = CaptureSetting());
+  void configure_capture_setting(const CaptureSetting & capture_setting = CaptureSetting());
 
-  inline const CaptureSetting & get_capture_setting() const;
+  const CaptureSetting & get_capture_setting() const;
 
 private:
   rclcpp::Publisher<CaptureSettingMsg>::SharedPtr capture_setting_event_publisher;
@@ -54,63 +55,6 @@ private:
 
   CaptureSetting current_capture_setting;
 };
-
-CaptureSettingProvider::CaptureSettingProvider(
-  rclcpp::Node::SharedPtr node, const CaptureSettingProvider::Options & options)
-: CameraNode(node, options)
-{
-  // Initialize the capture setting event publisher
-  {
-    capture_setting_event_publisher = get_node()->create_publisher<CaptureSettingMsg>(
-      get_camera_prefix() + CAPTURE_SETTING_EVENT_SUFFIX, 10);
-
-    RCLCPP_INFO_STREAM(
-      get_node()->get_logger(),
-      "Capture setting event publisher initialized on `" <<
-        capture_setting_event_publisher->get_topic_name() << "`!");
-  }
-
-  // Initialize the configure capture setting service
-  {
-    configure_capture_setting_service = get_node()->create_service<ConfigureCaptureSetting>(
-      get_camera_prefix() + CONFIGURE_CAPTURE_SETTING_SUFFIX,
-      [this](ConfigureCaptureSetting::Request::SharedPtr request,
-      ConfigureCaptureSetting::Response::SharedPtr response) {
-        // Configure capture setting if exist
-        if (request->capture_setting.size() > 0) {
-          configure_capture_setting((const CaptureSetting &)request->capture_setting.front());
-        }
-
-        response->capture_setting.push_back(get_capture_setting());
-      });
-
-    RCLCPP_INFO_STREAM(
-      get_node()->get_logger(),
-      "Configure capture setting service initialized on `" <<
-        configure_capture_setting_service->get_service_name() << "`!");
-  }
-
-  // Initial data fetching
-  ConfigureCaptureSetting();
-}
-
-CaptureSetting CaptureSettingProvider::on_configure_capture_setting(
-  const CaptureSetting & capture_setting)
-{
-  return capture_setting;
-}
-
-void CaptureSettingProvider::configure_capture_setting(const CaptureSetting & capture_setting)
-{
-  // Update with configured data
-  current_capture_setting.update_with(on_configure_capture_setting(capture_setting));
-  capture_setting_event_publisher->publish(get_capture_setting());
-}
-
-const CaptureSetting & CaptureSettingProvider::get_capture_setting() const
-{
-  return current_capture_setting;
-}
 
 }  // namespace shisen_cpp
